requirements.txt
asgiref==3.7.2
bootstrap4==0.1.0
crispy-bootstrap4==2023.1
Django==5.0.1
django-allauth==0.62.1
django-cors-headers==4.3.1
django-crispy-forms==2.1
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
joblib==1.3.2
pillow==10.3.0
PyJWT==2.8.0
sqlparse==0.4.4
typing_extensions==4.9.0
tzdata==2023.4
